frontend/requirements.txt
streamlit
requests
python-dotenv~=1.0.0
Pillow
pandas
numpy
streamlit-drawable-canvas
redis